two views implemented for css grid
