+++ src/components/LunarPhase/LunarPhase.css
@@ -1,15 +1,28 @@
-/* --------------------------------------------------------------------------------
-CSS Variables
--------------------------------------------------------------------------------- */
+/* ----- CSS VARIABLES ------ */
+
 :root {
     --moonlight: rgb(238, 238, 238);
     --moonshading: rgb(156, 156, 156);
     --moon-shadow: rgb(51, 51, 51);
 }
 
-/* --------------------------------------------------------------------------------
-  Moon
-  -------------------------------------------------------------------------------- */
+/* ----- CSS GRID ------ */
+.moon-clock {
+    grid-area: 1 / 1 / 3 / 2;
+}
+
+.moon-btns {
+    grid-area: 1 / 2 / 3 / 6;
+}
+
+.lunarnav-layout {
+    display: grid;
+    grid-template-columns: 200px repeat(4, 1fr);
+    grid-template-rows: 1fr 1 fr;
+    background-color: #000000;
+}
+
+/* ----- MOON CLOCK ------ */
 
 .moon {
     width: 150px;
